+++ src/main/java/com/owlmaddie/chat/ChatDataManager.java
@@ -487,7 +487,7 @@ public class ChatDataManager {
         } catch (Exception e) {
             String errorMessage = "Error saving `chatdata.json`. No CreatureChat chat history was saved! " + e.getMessage();
             LOGGER.error(errorMessage, e);
-            ServerPackets.sendMessageToAllOps(server, errorMessage);
+            ServerPackets.sendErrorToAllOps(server, errorMessage);
         }
     }